ModalDecisionLists

Welcome to the documentation for ModalDecisionLists.

ModalDecisionLists.AtomSearch

ModalDecisionLists.BeamSearch

ModalDecisionLists.RandSearch

ModalDecisionLists.SearchMethod

ModalDecisionLists.sequentialcovering

Modules = [ModalDecisionLists, ModalDecisionLists.MLJInterface, ModalDecisionLists.experimentals]

TODO continue this example

ModalDecisionLists.sequentialcovering — Function

function sequentialcovering(
    X::AbstractLogiset,
    y::AbstractVector{<:CLabel},
    w::Union{Nothing,AbstractVector{U},Symbol} = default_weights(length(y));
    kwargs...
)::DecisionList where {U<:Real}

Learn a decision list on an logiset X with labels y and weights w following the classic sequential covering learning scheme. This involves iteratively learning a single rule, and removing the newly covered instances.

Keyword Arguments

searchmethod::SearchMethod: The search method for finding single rules (see SearchMethod);

loss_function::Function = soleentropy is the function that assigns a score to each partial solution.

max_infogain_ratio::Real=1.0: constrains the maximum information gain for anantecedent with respect to the uncovered training set. Its value is bounded between 0 and 1.

default_alphabet::Union{Nothing,AbstractAlphabet}=nothing offers the flexibility to define a tailored alphabet upon which antecedents generation occurs.

discretizedomain::Bool=false: discretizes continuous variables by identifying optimal cut points

significance_alpha::Union{Real,Nothing}=0.0 is the significant alpha

min_rule_coverage::Union{Nothing,Integer} = 1 specifies the minimum number of instances covered by each rule.

max_rule_length::Union{Nothing,Integer} = nothing specifies the maximum length allowed for a rule in the search algorithm.

max_rulebase_length::Union{Nothing,Integer} is the maximum length of the rulebase;

suppress_parity_warning::Bool if true, suppresses parity warnings.

Any additional keyword argument will be imputed to the searchmethod, replacing its original value.

Examples

julia> X = PropositionalLogiset(iris_dataframe);

julia> y = Vector{CLabel}(iris_labels);

julia> sequentialcovering(X, y)
▣
├[1/22]┐(:sepal_length ≤ 4.8)
│└ setosa
├[2/22]┐(:sepal_length ≥ 7.1)
│└ virginica
├[3/22]┐(:sepal_length ≥ 7.0)
│└ versicolor
├[4/22]┐(:sepal_width ≤ 2.0)
│└ versicolor
├[5/22]┐(:sepal_width ≥ 3.5)
│└ setosa
├[6/22]┐(:petal_length ≤ 1.7)
│└ setosa
├[7/22]┐(:petal_length ≤ 4.4)
│└ versicolor
├[8/22]┐(:sepal_length ≤ 4.9)
│└ virginica
├[9/22]┐(:sepal_length ≤ 5.4)
│└ versicolor
├[10/22]┐(:petal_length ≤ 4.7)
│└ versicolor
├[11/22]┐(:sepal_length ≤ 5.8)
│└ virginica
├[12/22]┐(:sepal_width ≤ 2.2)
│└ virginica
├[13/22]┐(:sepal_width ≥ 3.3)
│└ virginica
├[14/22]┐(:petal_length ≥ 5.2)
│└ virginica
├[15/22]┐(:petal_width ≤ 1.4)
│└ versicolor
├[16/22]┐(:petal_width ≥ 1.9)
│└ virginica
├[17/22]┐(:sepal_length ≥ 6.7)
│└ versicolor
├[18/22]┐(:sepal_width ≤ 2.5)
│└ versicolor
├[19/22]┐(:sepal_length ≥ 6.1)
│└ virginica
├[20/22]┐(:sepal_width ≤ 2.7)
│└ versicolor
├[21/22]┐(:sepal_length ≥ 6.0)
│└ virginica
├[22/22]┐(:sepal_width ≤ 3.0)
│└ virginica
└✘ versicolor

See also SearchMethod, BeamSearch, PropositionalLogiset, DecisionList.

ModalDecisionLists.SearchMethod — Type

    SearchMethod

Abstract type for all search methods to be used in sequentialcovering.

Any search method implements a findbestantecedent method.

See also findbestantecedent, BeamSearch, RandSearch.

ModalDecisionLists.AtomSearch — Type

    AtomSearch

Method for conjunctions search to be used with BeamSearch, where the conjunctions are restricted to atomic conditions. This approach precisely implements the CN2 algorithm.

ModalDecisionLists.BeamSearch — Type

Search method to be used in sequentialcovering that explores the solution space selectively, maintaining a restricted set of partial solutions (the "beam") at each step.

The beam is dynamically updated to include the most promising solutions, allowing for efficient exploration of the solution space without examining all possibilities.

Keyword Arguments

conjuncts_search_method::SearchMethod=AtomSearch(): Defines the heuristic method by which possible conjuncts are generated during the beam search.

beam_width::Integer=3 is the width of the beam, i.e., the maximum number of partial solutions to maintain during the search.

See also sequentialcovering, SearchMethod, AtomSearch, RandSearch, specializeantecedents.

ModalDecisionLists.RandSearch — Type

RandSearch (`SoleLogics.randformula`)

Search method to be used in sequentialcovering that explores the solutions space employing stochastic sampling strategies.

Keyword Arguments

cardinality::Integer=25: Defines the number of formulas generated during the search for a single rule. A higher cardinality increases the probability of finding an antecedent that better fits the data.

operators::AbstractVector=[NEGATION, CONJUNCTION, DISJUNCTION]: Represents the set of logical operators used in the generation of formulas.

syntaxheight::Integer=2: Defines the maximum height of the syntactic tree representing a generated formula.

rng::AbstractRNG=Random.GLOBAL_RNG: Specifies the random number generator to be used in the generation of formulas. By default, it uses the global random number generator.

atompicking_mode::Symbol=:uniform: Determines the probability distribution of MetaConditions when generating formulas. It can impose a :uniform distribution over the MetaConditions or a :weighted distribution based on the length of the thresholding values of each MetaCondition.

subalphabets_weights::Union{AbstractWeights,AbstractVector{<:Real},Nothing}=nothing: Allows biasing the probability distribution of each MetaCondition through a vector of real weights between 0 and 1.

See also sequentialcovering, SearchMethod, BeamSearch, specializeantecedents.