ModalDecisionLists

Welcome to the documentation for ModalDecisionLists.

ModalDecisionLists.SearchMethod

ModalDecisionLists.sequentialcovering

Modules = [ModalDecisionLists, ModalDecisionLists.MLJInterface, ModalDecisionLists.experimentals]

TODO continue this example

ModalDecisionLists.sequentialcovering — Function

function sequentialcovering(
    X::AbstractLogiset,
    y::AbstractVector{<:CLabel},
    w::Union{Nothing,AbstractVector{U},Symbol} = default_weights(length(y));
    kwargs...
)::DecisionList where {U<:Real}

Learn a decision list on an logiset X with labels y and weights w following the classic sequential covering learning scheme. This involves iteratively learning a single rule, and removing the newly covered instances.

Keyword Arguments

searchmethod::SearchMethod: The search method for finding single rules (see SearchMethod);

max_rulebase_length::Union{Nothing,Integer} is the maximum length of the rulebase;

suppress_parity_warning::Bool if true, suppresses parity warnings.

unorderedstrategy::Bool: TODO @Edo explain

min_rule_coverage::Integer:

Any additional keyword argument will be imputed to the searchmethod, replacing its original value.

Examples

julia> X = PropositionalLogiset(iris_dataframe);

julia> y = Vector{CLabel}(iris_labels);

julia> sequentialcovering(X, y)
▣
├[1/22]┐(:sepal_length ≤ 4.8)
│└ setosa
├[2/22]┐(:sepal_length ≥ 7.1)
│└ virginica
├[3/22]┐(:sepal_length ≥ 7.0)
│└ versicolor
├[4/22]┐(:sepal_width ≤ 2.0)
│└ versicolor
├[5/22]┐(:sepal_width ≥ 3.5)
│└ setosa
├[6/22]┐(:petal_length ≤ 1.7)
│└ setosa
├[7/22]┐(:petal_length ≤ 4.4)
│└ versicolor
├[8/22]┐(:sepal_length ≤ 4.9)
│└ virginica
├[9/22]┐(:sepal_length ≤ 5.4)
│└ versicolor
├[10/22]┐(:petal_length ≤ 4.7)
│└ versicolor
├[11/22]┐(:sepal_length ≤ 5.8)
│└ virginica
├[12/22]┐(:sepal_width ≤ 2.2)
│└ virginica
├[13/22]┐(:sepal_width ≥ 3.3)
│└ virginica
├[14/22]┐(:petal_length ≥ 5.2)
│└ virginica
├[15/22]┐(:petal_width ≤ 1.4)
│└ versicolor
├[16/22]┐(:petal_width ≥ 1.9)
│└ virginica
├[17/22]┐(:sepal_length ≥ 6.7)
│└ versicolor
├[18/22]┐(:sepal_width ≤ 2.5)
│└ versicolor
├[19/22]┐(:sepal_length ≥ 6.1)
│└ virginica
├[20/22]┐(:sepal_width ≤ 2.7)
│└ versicolor
├[21/22]┐(:sepal_length ≥ 6.0)
│└ virginica
├[22/22]┐(:sepal_width ≤ 3.0)
│└ virginica
└✘ versicolor

See also SearchMethod, BeamSearch, PropositionalLogiset, DecisionList.

ModalDecisionLists.SearchMethod — Type

    SearchMethod

Abstract type for all search methods to be used in sequentialcovering.

Any search method implements a findbestantecedent method.

See also findbestantecedent, BeamSearch, RandSearch.

Missing docstring.

Missing docstring for AtomSearch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for BeamSearch. Check Documenter's build log for details.

Missing docstring.

Missing docstring for RandSearch. Check Documenter's build log for details.